Callbacks script
This script is called called by FPPD, it should initally return what
types of things we want to react against.

When those items play or run then the this script gets called with FPPD
supplying info about the item.

This is used to update your now playing sequence on the website