Added ability to specify Start and End events
Added ability to specify Start and End events, useful if you are using
events to run a looping effect (tune to sign) or other things.
Events are only used for the Juke Box playlist where we may never return
to the scheduled playlist.
During most voted playlists, we always return and play the schduled
playlist through